Client :  Refactor Settings Table into Entity Table. Created new components -> blocks folder. This is meant for bigger components that import other components / have a bit more logic. Widgets should then be used for small reusable components.
