+++ client/src/components/blocks/entity-table/widgets/DeleteDialog.vue
