+++ client/src/components/blocks/entity-table/widgets/PageTitle.vue
