+++ client/src/components/blocks/entity-table/widgets/StatusDialog.vue
