+++ client/src/components/layout/pages/BasicSettingsPage.vue
@@ -45,9 +45,9 @@
 import { onBeforeUnmount, onBeforeMount } from "vue";
 import { useAppBarStore, useLayoutStore } from "@/stores/store";
 import { translate } from "@/utils/generalUtils.ts";
-import PageTitle from "@/components/views/seb-server/settings-navigation/widgets/PageTitle.vue";
 import HorizontalDividerLine from "@/components/layout/pages/widgets/HorizontalDividerLine.vue";
-import SettingsNavigation from "@/components/layout/pages/widgets/SettingsNavigation.vue";
+import SettingsNavigation from "@/components/layout/pages/blocks/SettingsNavigation.vue";
+import PageTitle from "@/components/blocks/entity-table/widgets/PageTitle.vue";
 
 const layoutStore = useLayoutStore();
 const appBarStore = useAppBarStore();
