+++ client/src/components/layout/pages/widgets/HorizontalDividerLine.vue
@@ -1,7 +1,4 @@
 <script setup lang="ts"></script>
 <template>
-    <v-divider
-        class="custom-divider mx-10 my-4 mt-4"
-        data-testid="userAccounts-divider-top"
-    />
+    <v-divider class="mx-10 my-4 mt-4" />
 </template>
