+++ client/src/components/layout/pages/widgets/SearchSection.vue
