+++ client/src/components/views/seb-server/connection-configuration/connection-configurations/ConnectionConfigurations.vue
@@ -35,7 +35,7 @@
                         <div v-else-if="deleteError">{{ deleteError }}</div>
                         <div v-else-if="statusError">{{ statusError }}</div>
 
-                        <SettingsTable
+                        <EntityTable
                             :headers="connectionConfigurationTableHeaders"
                             :items="tableData?.content ?? []"
                             :total-items="totalItems"
@@ -69,10 +69,10 @@
 import { computed, ref, watch } from "vue";
 import BasicSettingsPage from "@/components/layout/pages/BasicSettingsPage.vue";
 import { getRouteName } from "@/router/routeNames.ts";
-import SearchSection from "@/components/views/seb-server/settings-navigation/components/SearchSection.vue";
+import SearchSection from "@/components/layout/pages/widgets/SearchSection.vue";
 import FiltersBar from "@/components/blocks/filters/FiltersBar.vue";
-import SettingsTable from "@/components/views/seb-server/settings-navigation/components/SettingsTable/SettingsTable.vue";
-import { useUrlSettingsTable } from "@/components/views/seb-server/settings-navigation/components/SettingsTable/composables/useUrlSettingsTable.ts";
+import EntityTable from "@/components/blocks/entity-table/EntityTable.vue";
+import { useUrlTableState } from "@/components/blocks/entity-table/composables/useUrlTableState.ts";
 import { useConnectionConfigurationsTableHeaders } from "@/components/views/seb-server/connection-configuration/connection-configurations/composables/useConnectionConfigurationsTableHeaders.ts";
 import { useConnectionConfigurationsFilters } from "@/components/views/seb-server/connection-configuration/connection-configurations/composables/useConnectionConfigurationsFilters.ts";
 import { STATUS_FILTER_KEY } from "@/components/blocks/filters/statusFilterSection.ts";
@@ -80,7 +80,7 @@ import { INSTITUTION_FILTER_KEY } from "@/components/blocks/filters/useInstituti
 import { useConnectionConfigurations } from "@/components/views/seb-server/connection-configuration/connection-configurations/api/useConnectionConfigurations.ts";
 import { useDeleteConnectionConfiguration } from "@/components/views/seb-server/connection-configuration/connection-configurations/api/useDeleteConnectionConfiguration.ts";
 import { useToggleConnectionConfigurationStatus } from "@/components/views/seb-server/connection-configuration/connection-configurations/api/useToggleConnectionConfigurationStatus.ts";
-import { useSettingsTableCellFormatters } from "@/components/views/seb-server/settings-navigation/components/SettingsTable/composables/useSettingsTableCellFormatters.ts";
+import { useTableCellFormatters } from "@/components/blocks/entity-table/composables/useTableCellFormatters.ts";
 import type { ConnectionConfigurations } from "@/models/seb-server/connectionConfiguration.ts";
 import AddButton from "@/components/widgets/AddButton.vue";
 
@@ -101,7 +101,7 @@ const {
     onClearSearch,
     setFilters,
     resetFilters,
-} = useUrlSettingsTable(tableData, async () => {
+} = useUrlTableState(tableData, async () => {
     await fetchConnectionConfigurations();
 }, [STATUS_FILTER_KEY, INSTITUTION_FILTER_KEY]);
 
@@ -144,7 +144,7 @@ const {
     error: statusError,
 } = useToggleConnectionConfigurationStatus(tableData);
 
-const { cellFormatters } = useSettingsTableCellFormatters({
+const { cellFormatters } = useTableCellFormatters({
     headers: connectionConfigurationTableHeaders,
 });
 </script>
